% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rnaseqTMLEout}
\alias{rnaseqTMLEout}
\title{Results obtained from running biomarkertmle on simulated RNA-Seq data}
\format{A \code{biotmle} object containing the results of running
        \code{biomarkertmle}.
\describe{
  These results are included here for the sake of making the vignettes build
  more quickly. The user will likely not benefit from using this data set.
}}
\usage{
rnaseqTMLEout
}
\value{
A \code{biotmle} object containing results from \code{biomarkertmle}.
}
\description{
Example results obtained from running the TMLE-based estimation procedure on
next-generation sequencing (count) data.
}
\keyword{datasets}
